split exposure_by_admin_region into 3 rules
