+++ workflow/rules/exposure/storm_damages.smk
@@ -283,27 +283,60 @@ def exposure_by_storm_for_country_for_storm_set(wildcards):
     )
 
 
+rule concat_exposure_by_event:
+    """
+    Take per-event exposure files with per-edge rows and concatenate into a single file.
+    """
+    input:
+        exposure_by_event = exposure_by_storm_for_country_for_storm_set,
+    # read exposure files in parallel
+    # use all available cores to block rest of execution (maximise memory available)
+    threads: workflow.cores
+    output:
+        concatenated = "{OUTPUT_DIR}/power/by_country/{COUNTRY_ISO_A3}/exposure/{STORM_SET}/exposure_by_event.parquet"
+    script:
+        "../../scripts/exposure/concat_grid_exposure.py"
+
+"""
+Test with:
+snakemake -n --cores 1 -- results/power/by_country/PRI/exposure/IBTrACS/exposure_by_event.parquet
+"""
+
+
 rule exposure_by_admin_region:
     """
     Calculate expected annual exposure at given admin level.
     """
     input:
         tracks = storm_tracks_file_from_storm_set,
-        exposure = exposure_by_storm_for_country_for_storm_set,
+        exposure_by_edge_by_event = rules.concat_exposure_by_event.output.concatenated,
         admin_areas = "{OUTPUT_DIR}/input/admin-boundaries/{ADMIN_SLUG}.geoparquet",
         grid_edges = rules.create_power_network.output.edges,
-    threads: 8  # read exposure files in parallel
     output:
         total_exposure_by_region = "{OUTPUT_DIR}/power/by_country/{COUNTRY_ISO_A3}/exposure/{STORM_SET}/{ADMIN_SLUG}.geoparquet",
-        # all outputs must have same wildcards, so country_event_distributions contains ADMIN_SLUG
-        # despite country_event_distributions being the same for every ADMIN_SLUG
-        country_event_distributions = directory("{OUTPUT_DIR}/power/by_country/{COUNTRY_ISO_A3}/exposure/{STORM_SET}/{ADMIN_SLUG}/country_event_distribution/")
     script:
         "../../scripts/exposure/grid_exposure_by_admin_region.py"
 
 """
 Test with:
 snakemake -c1 -- results/power/by_country/PRI/exposure/IBTrACS/admin-level-1.geoparquet
+"""
+
+
+rule plot_event_exposure_distributions_for_country:
+    """
+    Plot the exposure distribution (over events) for a given country and storm set.
+    """
+    input:
+        exposure_by_event = rules.concat_exposure_by_event.output.concatenated
+    output:
+        country_event_distributions = directory("{OUTPUT_DIR}/power/by_country/{COUNTRY_ISO_A3}/exposure/{STORM_SET}/event_distributions/")
+    script:
+        "../../scripts/exposure/plot_exposure_distributions.py"
+
+"""
+Test with:
+snakemake -c1 -- results/power/by_country/PRI/exposure/IBTrACS/country_event_distribution
 """
